refactor: getDistanceOfPin 훅으로 분리
